requirements.txt
numpy>=1.26.0
pandas>=2.2.0
duckdb>=1.0.0
pydantic>=2.6.0
openai>=1.30.0
openpyxl>=3.1.0
requests>=2.31.0

# Data contracts and validation
pandera>=0.20.0

# Optimization and advanced modeling
cvxpy>=1.5.0
scikit-learn>=1.4.0
rapidfuzz>=3.9.0
PyYAML>=6.0.1

# Alternate dataframe/IO engines
polars>=1.0.0
pyarrow>=16.0.0

# Service and orchestration
fastapi>=0.111.0
uvicorn>=0.30.0
prefect>=3.0.0

# Dev quality tooling
ruff>=0.5.0
mypy>=1.10.0
bandit>=1.7.8
